# swagger.yaml
swagger: '2.0'
info:
  version: '1.0'
  title: "Unisurvey"
  description: UniSurvey is an utility web service with the focus on improving students life with surveys and in-campus functionalities.
  license:
    name: MIT
    url: https://github.com/apiaryio/polls-api/blob/master/LICENSE
host: polls.apiblueprint.org
basePath: /api/
schemes:
- https
consumes:
- application/json
produces:
- application/json
paths:
  /student/signup:
    post:
      tags: [Student]
      summary: register a new student
      description: the student is created after some validation checks.
      responses:
           '409':
               description: 'email already in use: duplicate student'
           '400':
               description: 'bad request'
           '500':
               description: 'database error'
           '201':
               description: 'student successfully created'
  /students/login:
    post:
       tags: [Student]
       summary: log into an existing student
       description: the student is logged in after validation of his email and password. A JWT token is returned.
           
       responses:
           '400':
               description: 'Bad request'
               
           '404':
               description: 'email not found'
               
           '401':
               description: 'wrong password: failed authentication'
           '200':
               description: 'user logged in'
...
  /students/token:
    post:
        tags: [Student]
        summary: returns a new fresh access token
        description: the student is given another fresh JWT token access token after having recognized his JWT refresh token.
        responses:
            '403':
                description: 'forbidden'
            '401':
                description: 'unauthorized, unknown refresh token'
            '200':
                description: 'new refresh token'
  /students/:id:
   patch:
       tags: [Student]
       summary: update an existing student
       description: the student changes some of his personal datas like name, surname, gender and password.
       responses:
           '404':
               description: 'student not found'
               
           '500':
               description: 'database internal error'
               
           '202':
               description: 'student information updated succesfully'
  /classroomsBooking/:id:
    patch:
        tags: [Student]
        summary: book a classroom if available
        description: the student books a classroom if available, otherwise that option is negated
        responses:
            '404':
                description: 'classoom not found'

            '500':
                description: 'database internal error'

            '202':
                description: 'classroom booked succesfully'
...